remove older differs
