+++ lib/dyph3.rb
@@ -8,7 +8,6 @@ require "dyph3/outcome/resolved"
 require "dyph3/outcome/conflicted"
 
 require "dyph3/support/diff3"
-require "dyph3/support/diff3_beta"
 
 require "dyph3/support/collater"
 require "dyph3/support/merger"
@@ -16,7 +15,6 @@ require "dyph3/support/sanity_check"
 require "dyph3/support/assign_action"
 
 require "dyph3/two_way_differs/heckel_diff"
-require "dyph3/two_way_differs/original_heckel_diff"
 
 require "dyph3/two_way_differs/output_converter"
